Making containers persist until all objects associated with it have been deleted.
This is done with the help of object counter increments and decrements.
Containers are deleted only after a call from task_delete because the current assumption
ensures that the object delete cannot be called without a task, i.e. a task has to exist
for an object delete to be performed.